+++ lib/src/scroll_controller_extension.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
 
+/// Extension on [BuildContext] to provide quick access to [ScrollController] properties.
 extension ScrollControllerExtension on BuildContext {
-  ScrollController get primaryScrollController => PrimaryScrollController.of(this);
+  /// The primary [ScrollController] for the current [BuildContext].
+  ScrollController get primaryScrollController =>
+      PrimaryScrollController.of(this);
 }