+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.1
+
+* Removed unnecessary platform folders (android, ios, linux, macos, windows).
+* Documented and improved `example/example.dart` with more use cases.
+* Generated and fixed test cases in `test/simple_build_context_test.dart`.
+* Updated to the latest version of Flutter and improved code documentation.
+
 ## 1.0.0
 
 * TODO: Describe initial release.
@@ -20,5 +27,5 @@
 
 
 ## 1.0.5
-## 1.0.6
+
 * Upgrade SDK